areas_legales_crud_V1
falta manejar el crud con un campo extra para las imagenes

+++ consultoriaHB/src/app/components/admin/dynamic/legal-areas/list/list.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Output, EventEmitter} from '@angular/core';
 import { LegalAreasService} from '../../../../../services/legal_areas/legal-areas.service';
 import { AreaLegal} from '../../../../../models/areaLegal';
 import { map } from 'rxjs/operators';
@@ -11,6 +11,9 @@ import { Observable } from 'rxjs';
   styleUrls: ['./list.component.css']
 })
 export class ListComponent implements OnInit {
+
+  @Output() dataItemArea = new EventEmitter<boolean>();
+
   i = 1;
   listArea: AreaLegal[] = [];
   list: AreaLegal[]=[];//We are using this variable for showing data instead of calling FireBase database again
@@ -26,18 +29,20 @@ export class ListComponent implements OnInit {
   //  this.listArea[0] = {$id:"a",titulo:"sda",contenido:"asd",fecha:"asd",hora:"asd"};
     this.list=this.listArea;
   }
-  updateArea(id:string,data: AreaLegal) {
-    /*this.infoService
-      .updateInfo(id, data)
-      .catch(err => console.log(err));
-      
-      */
+
+  openFormEdit(data: AreaLegal) {
+    this.dataItemArea.emit(true);
+    this.areaService.selectedAreaLegal = Object.assign({}, data);   
   }
   deleteArea(id:string) {
-    console.log(id);
-    this.areaService.deleteArea(id).catch(
-      err => console.log(err)
-    );
+    if (confirm("¿Esta seguro de quere eliminar a este elemento?")){
+      this.areaService.deleteArea(id).catch(
+        err => console.log(err)
+      );
+    }
+   // this.areaService.deleteArea(id).catch(
+   //   err => console.log(err)
+   // );
   }
 
   getAreaList() {
